Move ErrorMessage to flat structure
