+++ packages/client/src/views/Unlock/Unlock.tsx
@@ -19,7 +19,7 @@ import styled from '@client/styledComponents'
 import { SECURITY_PIN_EXPIRED_AT } from '@client/utils/constants'
 import { IUserDetails, getUserName } from '@client/utils/userUtils'
 import { pinValidator } from '@client/views/Unlock/ComparePINs'
-import { ErrorMessage } from '@opencrvs/components/lib/forms'
+import { ErrorMessage } from '@opencrvs/components/lib/ErrorMessage'
 import { Logout } from '@opencrvs/components/lib/icons'
 import { PINKeypad } from '@opencrvs/components/lib/interface'
 import { GQLHumanName } from '@opencrvs/gateway/src/graphql/schema'
